+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Implemented Usernames endpoint in individual member reports
 - Added USN's in members page
 - Fixed the Table for Mobile View
+- Added PWA Support
 
 ## v0.3.0